move storage classes
